Add '(optional)' to 'Real name' field on RequestAccount
While the CreateAccount page lists that field as "Real name (optional)",
even with the additional description "Real name is optional. If you
choose to provide it, this will be used for giving the user attribution
for their work.", on the RequestAccount page this is not marked at all,
leading users to think they're obligated to put their real name there.

Make it visible that it's actually optional by updating the string.

Change-Id: If789ca823925c8b5a83d1e70f38808216fbcadca