Replace use of deprecated Xml::radio()
Also use the namespaced reference to the Xml class in the three files that
were previously using it via the deprecated global namespace.

The other uses of the class should also be migrated, but this is the last
use in public gerrit of this method, so landing this now and back-porting
will allow us to deprecate it that bit earlier.

Bug: T341775
Change-Id: Ia23a2346f6990414b14c3b0424aecba6798eeb65